+++ infrastructure/esmf/API.F90
@@ -2,10 +2,71 @@
 ! Public API of esmf/ leaf modules exposed to external consumers.
 module mapl_esmf_api
 
-   use mapl_esmf_internal
+   ! Core ESMF utilities
+   use mapl_ESMF_Interfaces_mod
+   use mapl_ESMF_Utilities_mod
+   use mapl_ESMF_Time_Utilities_mod
+   use mapl_ESMF_Subset_mod
+   use mapl_ESMF_HConfigUtilities_mod
 
-   use mapl_Comms_mod, only: arraygather => array_gather
-   use mapl_Comms_mod, only: arrayscatter => array_scatter
+   ! Info / metadata utilities
+   use mapl_InfoUtilities_mod
+
+   ! Ungridded dimensions
+   use mapl_UngriddedDim_mod
+   use mapl_UngriddedDims_mod
+   use mapl_UngriddedDimVector_mod
+
+   ! VM
+   use mapl_vm_mod, only: mapl_AmIRoot, mapl_AmIPet, mapl_Barrier
+
+   ! Bounds / grid utilities
+   use mapl_LU_Bound_mod
+   use mapl_HorizontalDimsSpec_mod
+   use mapl_DistGridGet_mod
+   use mapl_FieldPointerUtilities_mod
+
+   ! HConfig
+   use mapl_HConfigAs_mod, only: &
+        mapl_HConfigAsItemType => HConfigAsItemType, &
+        mapl_HConfigAsStateIntent => HConfigAsStateIntent, &
+        mapl_HConfigAsTime => HConfigAsTime, &
+        mapl_HConfigAsTimeInterval => HConfigAsTimeInterval, &
+        mapl_HConfigAsTimeRange => HConfigAsTimeRange, &
+        mapl_HConfigAsStringVector => HConfigAsStringVector
+   use mapl_HConfigAs_mod
+   use mapl_HConfigUtilities_mod
+   use mapl_get_hconfig_mod
+   use mapl_hconfig_get_mod
+   use mapl_hconfig_params_mod
+   use mapl_generalized_equality_mod
+
+   ! Alarm
+   use mapl_SimpleAlarm_mod
+
+   ! State item
+   use mapl_StateItem_mod
+
+   ! Comms
+   use mapl_Comms_mod, only: ROOT_PROCESS_ID
+   use mapl_Comms_mod, only: &
+        MAPL_Am_I_Root => am_i_root, &
+        MAPL_Am_I_Rank => am_i_rank, &
+        MAPL_NPES => num_pes, &
+        MAPL_CommsSend => comms_send, &
+        MAPL_CommsRecv => comms_recv, &
+        MAPL_CommsSendRecv => comms_sendrecv, &
+        MAPL_CommsGatherV => comms_gatherv, &
+        MAPL_CommsScatterV => comms_scatterv, &
+        MAPL_CommsAllGather => comms_allgather, &
+        MAPL_CommsAllGatherV => comms_allgatherv, &
+        MAPL_ArrayGather => array_gather, &
+        MAPL_ArrayScatter => array_scatter, &
+        MAPL_CommsAllReduceMin => comms_allreduce_min, &
+        MAPL_CommsAllReduceMax => comms_allreduce_max, &
+        MAPL_CommsAllReduceSum => comms_allreduce_sum
+   use mapl_Comms_mod, only: arraygather => array_gather, arrayscatter => array_scatter
+   use mapl_ShmemComms_mod
    use mapl_FieldPointerUtilities_mod, only: MAPL_AssignFptr => assign_fptr
    use mapl_FieldPointerUtilities_mod, only: FieldGetLocalElementCount
    use mapl_FieldPointerUtilities_mod, only: mapl_FieldClone => FieldClone
@@ -35,8 +96,8 @@ module mapl_esmf_api
    public :: MAPL_CommsAllGatherV
    public :: MAPL_ArrayGather
    public :: MAPL_ArrayScatter
-   public :: ArrayGather
-   public :: ArrayScatter
+   public :: arraygather
+   public :: arrayscatter
    public :: MAPL_CommsAllReduceMin
    public :: MAPL_CommsAllReduceMax
    public :: MAPL_CommsAllReduceSum